pyproject.toml
[tool.poetry]
name = "crypto-arbitrage"
version = "0.1.0"
description = ""
authors = ["bertek <81099969+albertoecf@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
ccxt = "^4.3.91"
python-dotenv = "^1.0.1"
aiokafka = "^0.11.0"
pydantic = "^2.8.2"
six = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
